+++ src/test/java/ru/l3r8y/complaint/BulkComplaintTest.java
@@ -21,30 +21,49 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
 package ru.l3r8y.complaint;
 
+import java.nio.file.Paths;
 import java.util.Collection;
-import java.util.stream.Collectors;
-import lombok.AllArgsConstructor;
+import java.util.Collections;
+import java.util.regex.Pattern;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Test;
 import ru.l3r8y.Complaint;
+import ru.l3r8y.parser.ParsedMethod;
 
 /**
- * The compound complaint. Merges all complaints into one.
+ * Test case for {@link BulkComplaint}.
  *
- * @since 0.1.0
+ * @since 0.1.4
  */
-@AllArgsConstructor
-public final class CompoundComplaint implements Complaint {
+class BulkComplaintTest {
 
     /**
-     * The complaints.
+     * Separator for test.
      */
-    private final Collection<? extends Complaint> complaints;
+    private static final Pattern SEP = Pattern.compile("<sep>");
 
-    @Override
-    public String message() {
-        return this.complaints.stream()
-            .map(Complaint::message)
-            .collect(Collectors.joining("\n", "\n", ""));
+    @Test
+    void mergesMessages() {
+        final Collection<Complaint> complaints = Collections.nCopies(
+            5,
+            new WrongMethodSignature(
+                new ParsedMethod(
+                    "ClassName",
+                    "myCoolMethod()",
+                    "{ return null; }",
+                    Paths.get("")
+                ),
+                "some cool explanation!<sep>"
+            )
+        );
+        MatcherAssert.assertThat(
+            "Length before equals length after",
+            BulkComplaintTest.SEP.split(new BulkComplaint(complaints).message()).length,
+            Matchers.equalTo(complaints.size())
+        );
     }
 }
